Remove npm decidim packages with dependencies from other decidim packages
* chore: remove interpackages dependencies

* chore: don't publish root package

* fix: add all dependencies directly to root package.json

* chore: read npm dependencies from root json

* chore: regenerate webpacker binstubs instead of reinstalling it

Rails installs webpacker when creating the app, so we were installing webpacker twice.

* refactor: improve npm dependencies installation semantics

* fix: apply suggestions

* chore: update npm lock files

* test: add missing tests for tasks

* chore: remove unused package from lock file

* test: fix tests broken by webpacker changes

+++ decidim-core/lib/tasks/decidim_webpacker_tasks.rake
@@ -55,12 +55,8 @@ namespace :decidim do
     end
 
     def install_decidim_npm
-      decidim_npm_packages.each do |type, package|
-        if type == :dev
-          system! "npm i -D #{package}"
-        else
-          system! "npm i #{package}"
-        end
+      decidim_npm_packages.each do |type, packages|
+        system! "npm i --save-#{type} #{packages.join(" ")}"
       end
     end
 
@@ -73,32 +69,55 @@ namespace :decidim do
     end
 
     def decidim_npm_packages
+      gem_path = unreleased_gem_path
+
+      if gem_path
+        package_spec = "./packages/%s"
+
+        # The packages folder needs to be copied to the application folder
+        # because the linked dependencies are not installed when packages
+        # are installed using file references outside the application root
+        # where the `package.json` is located at. For more information, see:
+        # https://github.com/npm/cli/issues/2339
+        FileUtils.rm_rf(rails_app_path.join("packages"))
+        FileUtils.cp_r(gem_path.join("packages"), rails_app_path)
+      else
+        package_spec = "@decidim/%s@~#{Decidim::GemManager.semver_friendly_version(decidim_gemspec.version.to_s)}"
+      end
+
+      local_npm_dependencies.transform_values { |names| names.map { |name| format(package_spec, name) } }
+    end
+
+    def unreleased_gem_path
       if decidim_gemspec.source.is_a?(Bundler::Source::Rubygems)
-        if released_version?
-          return {
-            dev: "@decidim/dev@~#{Decidim::GemManager.semver_friendly_version(decidim_gemspec.version.to_s)}",
-            prod: "@decidim/all@~#{Decidim::GemManager.semver_friendly_version(decidim_gemspec.version.to_s)}"
-          }
-        else
-          gem_path = Pathname(decidim_gemspec.full_gem_path)
-        end
+        return if released_version?
+
+        gem_path = Pathname(decidim_gemspec.full_gem_path)
       else
         gem_path = decidim_gemspec.source.path
         gem_path = Pathname(ENV["BUNDLE_GEMFILE"]).dirname.join(gem_path) if gem_path.relative?
       end
 
-      # The packages folder needs to be copied to the application folder
-      # because the linked dependencies are not installed when packages
-      # are installed using file references outside the application root
-      # where the `package.json` is located at. For more information, see:
-      # https://github.com/npm/cli/issues/2339
-      FileUtils.rm_rf(rails_app_path.join("packages"))
-      FileUtils.cp_r(gem_path.join("packages"), rails_app_path)
+      gem_path
+    end
 
-      {
-        dev: "./packages/dev",
-        prod: "./packages/all"
-      }
+    def local_npm_dependencies
+      @local_npm_dependencies ||= begin
+        package_json = JSON.parse(File.read(decidim_path.join("package.json")))
+
+        {
+          prod: local_npm_dependencies_list(package_json["dependencies"]),
+          dev: local_npm_dependencies_list(package_json["devDependencies"])
+        }.freeze
+      end
+    end
+
+    def local_npm_dependencies_list(deps)
+      return [] unless deps
+
+      deps.values
+          .select { |ref| ref.starts_with?("file:packages/") }
+          .map { |ref| ref.delete_prefix("file:packages/") }
     end
 
     def decidim_path
@@ -151,8 +170,8 @@ namespace :decidim do
       File.write(file, contents)
     end
 
-    def system!(*args)
-      system(*args) || abort("\n== Command #{args} failed ==")
+    def system!(command)
+      system("cd #{rails_app_path} && #{command}") || abort("\n== Command #{args} failed ==")
     end
   end
 end
